moving stuff into the superclass
